# frontend-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "description": "A stable, machine-readable description of an API error.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "The common wire format for API errors.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "error",
          "requestId"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ReportingAcceptedResponse": {
        "properties": {
          "commandId": {
            "format": "uuid",
            "title": "Commandid",
            "type": "string"
          },
          "runId": {
            "title": "Runid",
            "type": "string"
          },
          "status": {
            "const": "accepted",
            "default": "accepted",
            "title": "Status",
            "type": "string"
          },
          "taskId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taskid"
          }
        },
        "required": [
          "commandId",
          "runId"
        ],
        "title": "ReportingAcceptedResponse",
        "type": "object"
      },
      "ReportingDecisionRequest": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "enum": [
              "supplement",
              "draft",
              "skip",
              "stop"
            ],
            "title": "Action",
            "type": "string"
          },
          "supplements": {
            "items": {
              "$ref": "#/components/schemas/UserSupplement"
            },
            "title": "Supplements",
            "type": "array"
          }
        },
        "required": [
          "action"
        ],
        "title": "ReportingDecisionRequest",
        "type": "object"
      },
      "ReportingListResponse": {
        "properties": {
          "runs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "runs"
        ],
        "title": "ReportingListResponse",
        "type": "object"
      },
      "ReportingResumeRequest": {
        "additionalProperties": false,
        "properties": {
          "maxProviderAttempts": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxproviderattempts"
          },
          "maxTotalTokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxtotaltokens"
          },
          "supplements": {
            "items": {
              "$ref": "#/components/schemas/UserSupplement"
            },
            "title": "Supplements",
            "type": "array"
          }
        },
        "title": "ReportingResumeRequest",
        "type": "object"
      },
      "ReportingRevisionRequest": {
        "additionalProperties": false,
        "properties": {
          "baselineVersionId": {
            "title": "Baselineversionid",
            "type": "string"
          },
          "feedback": {
            "title": "Feedback",
            "type": "string"
          },
          "maxProviderAttempts": {
            "default": 40,
            "title": "Maxproviderattempts",
            "type": "integer"
          },
          "maxTotalTokens": {
            "default": 400000,
            "title": "Maxtotaltokens",
            "type": "integer"
          },
          "promoteSkillId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promoteskillid"
          },
          "promoteToSkill": {
            "default": false,
            "title": "Promotetoskill",
            "type": "boolean"
          },
          "targetClaimIds": {
            "items": {
              "type": "string"
            },
            "title": "Targetclaimids",
            "type": "array"
          },
          "targetModuleIds": {
            "items": {
              "enum": [
                "2.1",
                "2.2",
                "2.3",
                "2.4",
                "2.5"
              ],
              "type": "string"
            },
            "title": "Targetmoduleids",
            "type": "array"
          },
          "targetSubmoduleIds": {
            "items": {
              "type": "string"
            },
            "title": "Targetsubmoduleids",
            "type": "array"
          },
          "userSupplements": {
            "items": {
              "$ref": "#/components/schemas/UserSupplement"
            },
            "title": "Usersupplements",
            "type": "array"
          }
        },
        "required": [
          "baselineVersionId",
          "feedback",
          "targetModuleIds"
        ],
        "title": "ReportingRevisionRequest",
        "type": "object"
      },
      "ReportingSnapshotResponse": {
        "properties": {
          "checkpoint": {
            "additionalProperties": true,
            "title": "Checkpoint",
            "type": "object"
          },
          "evidence": {
            "additionalProperties": true,
            "title": "Evidence",
            "type": "object"
          },
          "outputs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Outputs",
            "type": "array"
          },
          "revision": {
            "additionalProperties": true,
            "title": "Revision",
            "type": "object"
          },
          "run": {
            "additionalProperties": true,
            "title": "Run",
            "type": "object"
          },
          "state": {
            "additionalProperties": true,
            "title": "State",
            "type": "object"
          },
          "waitingInput": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Waitinginput",
            "type": "array"
          }
        },
        "required": [
          "run",
          "state",
          "waitingInput",
          "checkpoint",
          "evidence",
          "revision",
          "outputs"
        ],
        "title": "ReportingSnapshotResponse",
        "type": "object"
      },
      "ReportingStartRequest": {
        "additionalProperties": false,
        "properties": {
          "executionRequirements": {
            "items": {
              "type": "string"
            },
            "title": "Executionrequirements",
            "type": "array"
          },
          "instruction": {
            "minLength": 1,
            "title": "Instruction",
            "type": "string"
          },
          "maxProviderAttempts": {
            "default": 80,
            "title": "Maxproviderattempts",
            "type": "integer"
          },
          "maxTotalTokens": {
            "default": 800000,
            "title": "Maxtotaltokens",
            "type": "integer"
          },
          "missingEvidencePolicy": {
            "default": "draft",
            "enum": [
              "ask",
              "block",
              "skip",
              "draft"
            ],
            "title": "Missingevidencepolicy",
            "type": "string"
          },
          "operation": {
            "default": "full_report",
            "enum": [
              "distill_template_skill",
              "full_report",
              "module_report",
              "aggregate_existing",
              "render_existing"
            ],
            "title": "Operation",
            "type": "string"
          },
          "outputFilename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputfilename"
          },
          "sourceMarkdownRef": {
            "anyOf": [
              {
                "format": "path",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcemarkdownref"
          },
          "sourceModuleRefs": {
            "anyOf": [
              {
                "additionalProperties": {
                  "format": "path",
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcemodulerefs"
          },
          "targetModules": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetmodules"
          },
          "userSupplements": {
            "items": {
              "$ref": "#/components/schemas/UserSupplement"
            },
            "title": "Usersupplements",
            "type": "array"
          }
        },
        "required": [
          "instruction"
        ],
        "title": "ReportingStartRequest",
        "type": "object"
      },
...
      "UserSupplement": {
        "additionalProperties": false,
        "description": "One scoped, auditable user fact or instruction added to the current run.",
        "properties": {
          "content": {
            "description": "Exact user-confirmed fact or instruction without inferred additions.",
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "id": {
            "description": "Stable current-run supplement id used by supersedes.",
            "pattern": "^US-[A-Za-z0-9._-]+$",
            "title": "Id",
            "type": "string"
          },
          "scope": {
            "default": "run",
            "description": "Run-wide or explicitly targeted applicability boundary.",
            "enum": [
              "run",
              "module",
              "submodule",
              "claim",
              "final_section"
            ],
            "title": "Scope",
            "type": "string"
          },
          "stages": {
            "description": "Only workflow stages allowed to consume this supplement.",
            "items": {
              "enum": [
                "module_authoring",
                "module_review",
                "cross_review",
                "chief_edit",
                "final_review"
              ],
              "type": "string"
            },
            "title": "Stages",
            "type": "array"
          },
          "supersedes": {
            "description": "Earlier supplement ids replaced by this one while preserving audit history.",
            "items": {
              "type": "string"
            },
            "title": "Supersedes",
            "type": "array"
          },
          "target_ids": {
            "description": "Required module, submodule, Claim, or final-section ids outside run scope.",
            "items": {
              "type": "string"
            },
            "title": "Target Ids",
            "type": "array"
          }
        },
        "required": [
          "content"
        ],
        "title": "UserSupplement",
        "type": "object"
      }
    },
    "securitySchemes": {
      "ControlLeaseToken": {
        "in": "header",
        "name": "X-Control-Lease-Token",
        "type": "apiKey"
      },
      "SessionCookie": {
        "in": "cookie",
        "name": "manyselves_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/reporting/decisions/{decision_id}/resume": {
      "post": {
        "operationId": "post_api_v1_reporting_decisions_decision_id_resume",
        "parameters": [
          {
            "in": "path",
            "name": "decision_id",
            "required": true,
            "schema": {
              "title": "Decision Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportingDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportingAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "ControlLeaseToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Resume Decision"
      }
    },
    "/api/v1/reporting/revisions": {
      "post": {
        "operationId": "post_api_v1_reporting_revisions",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportingRevisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportingAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "ControlLeaseToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Revise"
      }
    },
    "/api/v1/reporting/runs": {
      "get": {
        "operationId": "get_api_v1_reporting_runs",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportingListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "List Runs"
      },
      "post": {
        "operationId": "post_api_v1_reporting_runs",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportingStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportingAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "ControlLeaseToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Start Run"
      }
    },
    "/api/v1/reporting/runs/{run_id}": {
      "get": {
        "operationId": "get_api_v1_reporting_runs_run_id",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportingSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "summary": "Get Run"
      }
    },
    "/api/v1/reporting/runs/{run_id}/cancel": {
      "post": {
        "operationId": "post_api_v1_reporting_runs_run_id_cancel",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportingAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "ControlLeaseToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Cancel Run"
      }
    },
    "/api/v1/reporting/runs/{run_id}/resume": {
      "post": {
        "operationId": "post_api_v1_reporting_runs_run_id_resume",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportingResumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportingAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Request validation failed"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "API error"
          }
        },
        "security": [
          {
            "ControlLeaseToken": [],
            "SessionCookie": []
          }
        ],
        "summary": "Resume Run"
      }
    },
...
  }
}